requirements.txt
blinker==1.9.0
click==8.1.8
confluent-kafka==2.9.0
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
psycopg2-binary==2.9.10
Werkzeug==3.1.3
